requirements.txt
streamlit
langchain==0.2.16
langchain-community==0.2.16
langchain-chroma
langchain-huggingface
langchain-groq
langchain-text-splitters
chromadb
sentence-transformers
newspaper3k
tiktoken
python-dotenv
lxml_html_clean